Bugfix for deadlock when locking the root object store without also locking SharedEditingContext
There is a potential race condition, when a rootObjectStore is locked, without prior locking the SharedEditingContext.

A sharedEditingContext is write locked, and then the (root)objectStore is locked in EOEditingContext.lockObjectStore. This method is called in EOCustomObject.readObject for example.
When the rootObjectStore is locked without also locking the sharedEditingContext, a deadlock may occur in globalIDForObject (or objectForGlobalID) when the EOEditingContexts implementation tries to lock the SharedEditingContext itself.

Thread 1:
Locks RootObjectStore.
Calls globalIDForObject -> locks SharedEditingContext.

Thread 2:
willRead calls EOEditingContext.lockObjextStore:
  -> locks SharedEditingContext
  -> locksRootObjectStore.

=> deadlock

This patch locks the sharedEditingContext prior to locking the rootObjectStore.