Add log4j xml configuration option.
Log4j has a time based rolling file appender, but discouriging its use. There is a another time based rolling file in log4j-extras, but this can only be configured by xml files, not system properties.
This patch introduces a log4j.configuration property, that can be used to read an additional configuration file. If the file name ends in .xml, than it is processed by log4j-extras DOMConfigurator, possible enabling the use of the rolling file appender.
It also moves evaluating the property er.extension.erxloggerclass to `ERXLogger.configureLoggingWithSystemProperties()`, because there are multiple places where this method is called, and the old implementation was only used in one place, effectively disabling the property.